+++ crates/alopex-sql/nim-sql-parser/src/lexer.nim
@@ -169,6 +169,11 @@ proc readString(lex: var Lexer): Token =
       value &= $c
   lex.makeToken(tkString, value, startLine, startCol)
 
+proc readQuotedIdent(lex: var Lexer): Token =
+  let token = lex.readString()
+  Token(kind: tkIdent, value: token.value, line: token.line, col: token.col,
+        endLine: token.endLine, endCol: token.endCol)
+
 proc readNumber(lex: var Lexer): Token =
   let startLine = lex.line
   let startCol = lex.col
@@ -209,8 +214,10 @@ proc nextToken*(lex: var Lexer): Token =
   let c = lex.peek()
 
   case c
-  of '\'', '"':
+  of '\'':
     return lex.readString()
+  of '"':
+    return lex.readQuotedIdent()
   of '0'..'9':
     return lex.readNumber()
   of 'a'..'z', 'A'..'Z', '_':